DDF-4166 Relocated spock-shaded to codice-test
