+++ catalog/admin/catalog-admin-poller-service/pom.xml
@@ -60,11 +60,6 @@
             <artifactId>action-core-api</artifactId>
         </dependency>
         <dependency>
-            <groupId>ddf.lib</groupId>
-            <artifactId>spock-shaded</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
             <groupId>ddf.admin.core</groupId>
             <artifactId>admin-core-impl</artifactId>
             <version>${project.version}</version>
