+++ catalog/core/catalog-core-api-impl/pom.xml
@@ -61,12 +61,6 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>ddf.lib</groupId>
-            <artifactId>spock-shaded</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
             <groupId>ddf.action.core</groupId>
             <artifactId>action-core-api</artifactId>
             <version>${project.version}</version>
