+++ catalog/core/catalog-core-standardframework/pom.xml
@@ -357,13 +357,6 @@
             <scope>provided</scope>
         </dependency>
 
-        <!--Spock dependencies-->
-        <dependency>
-            <groupId>ddf.lib</groupId>
-            <artifactId>spock-shaded</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
         <dependency>
             <groupId>org.awaitility</groupId>
             <artifactId>awaitility</artifactId>
