+++ catalog/core/catalog-core-validationparser/pom.xml
@@ -116,14 +116,6 @@
                 </exclusion>
             </exclusions>
         </dependency>
-
-        <!--Spock dependencies-->
-        <dependency>
-            <groupId>ddf.lib</groupId>
-            <artifactId>spock-shaded</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
     </dependencies>
 
     <build>
