+++ catalog/core/catalog-core-validator/pom.xml
@@ -52,13 +52,6 @@
             <artifactId>converter</artifactId>
             <version>${countryconverter.version}</version>
         </dependency>
-        <!--Spock dependencies-->
-        <dependency>
-            <groupId>ddf.lib</groupId>
-            <artifactId>spock-shaded</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
     </dependencies>
 
     <build>
