+++ catalog/plugin/catalog-plugin-gazetteer/pom.xml
@@ -46,13 +46,6 @@
       <artifactId>spatial-geocoding-api</artifactId>
       <version>${project.version}</version>
     </dependency>
-    <!-- Test dependencies -->
-    <dependency>
-      <groupId>ddf.lib</groupId>
-      <artifactId>spock-shaded</artifactId>
-      <version>${project.version}</version>
-      <scope>test</scope>
-    </dependency>
   </dependencies>
 
   <build>
