+++ catalog/transformer/catalog-transformer-pdf/pom.xml
@@ -93,14 +93,6 @@
             <artifactId>tika-parsers</artifactId>
             <scope>test</scope>
         </dependency>
-
-        <!--Spock dependencies-->
-        <dependency>
-            <groupId>ddf.lib</groupId>
-            <artifactId>spock-shaded</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
     </dependencies>
 
     <build>
