--- libs/spock-shaded/src/main/java/org/codice/spock/extension/DeFinalize.java
@@ -1,44 +1,0 @@
-/**
- * Copyright (c) Codice Foundation
- *
- * <p>This is free software: you can redistribute it and/or modify it under the terms of the GNU
- * Lesser General Public License as published by the Free Software Foundation, either version 3 of
- * the License, or any later version.
- *
- * <p>This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Lesser General Public License for more details. A copy of the GNU Lesser General Public
- * License is distributed along with this program and can be found at
- * <http://www.gnu.org/licenses/lgpl.html>.
- */
-package org.codice.spock.extension;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Inherited;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * This annotation is used to indicate to the {@link org.codice.spock.extension.builtin.DeFinalizer}
- * test runner, which classes or packages should be definalized.
- */
-@Inherited
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface DeFinalize {
-
-  /**
-   * Specifies the classes to de-finalize.
-   *
-   * @return the classes to de-finalize
-   */
-  public Class<?>[] value() default {};
-
-  /**
-   * Specifies the packages to de-finalize.
-   *
-   * @return the fully qualified names for the packages to de-finalize
-   */
-  public String[] packages() default {};
-}
