+++ platform/error/platform-error-page-injector/pom.xml
@@ -51,14 +51,6 @@
             <artifactId>platform-error-servlet</artifactId>
             <version>${project.version}</version>
         </dependency>
-
-        <!--Spock dependencies-->
-        <dependency>
-            <groupId>ddf.lib</groupId>
-            <artifactId>spock-shaded</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
     </dependencies>
 
     <build>
