+++ platform/landing-page/pom.xml
@@ -60,11 +60,6 @@
             <version>3.3.2</version>
         </dependency>
         <dependency>
-            <groupId>ddf.lib</groupId>
-            <artifactId>spock-shaded</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
             <groupId>com.github.jknack</groupId>
             <artifactId>handlebars</artifactId>
             <version>${jknack.handlebars.version}</version>
