+++ platform/security/idp/security-idp-plugin/security-idp-plugin-requestsubjectvalidator/pom.xml
@@ -32,14 +32,6 @@
       <artifactId>security-idp-plugin-api</artifactId>
       <version>${project.version}</version>
     </dependency>
-
-    <!--Spock dependencies-->
-    <dependency>
-      <groupId>ddf.lib</groupId>
-      <artifactId>spock-shaded</artifactId>
-      <version>${project.version}</version>
-      <scope>test</scope>
-    </dependency>
   </dependencies>
 
   <build>
