+++ platform/security/idp/security-idp-plugin/security-idp-plugin-subjectconfirmation/pom.xml
@@ -46,14 +46,6 @@
       <artifactId>platform-util-unavailableurls</artifactId>
       <version>${project.version}</version>
     </dependency>
-
-    <!--Spock dependencies-->
-    <dependency>
-      <groupId>ddf.lib</groupId>
-      <artifactId>spock-shaded</artifactId>
-      <version>${project.version}</version>
-      <scope>test</scope>
-    </dependency>
   </dependencies>
 
   <build>
