+++ platform/solr/solr-factory/pom.xml
@@ -101,12 +101,6 @@
         </dependency>
         <dependency>
             <groupId>ddf.lib</groupId>
-            <artifactId>spock-shaded</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>ddf.lib</groupId>
             <artifactId>failsafe-controller</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
