+++ platform/solr/solr-factory/src/test/groovy/org/codice/solr/factory/impl/UnavailableSolrClientSpec.groovy
@@ -15,7 +15,7 @@ package org.codice.solr.factory.impl
 
 import org.apache.solr.client.solrj.SolrClient
 import org.codice.solr.client.solrj.UnavailableSolrException
-import org.codice.spock.extension.Supplemental
+import org.codice.spock.Supplemental
 import spock.lang.Specification
 import spock.lang.Unroll
